% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySignature.R
\name{applySignature}
\alias{applySignature}
\title{Compute the Quantitative Scores and the Clinical Scores for each patient and for one or several gene signature(s)}
\usage{
applySignature(
  signatureUsed = "classification",
  distinguishingGenes = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{signatureUsed}{A character vector which tells rather to use the Classification Signature (if set to \code{classification}) or one or several custom gene signatures (if set to \code{custom}). The Classification Signature is read from the \verb{output > signatures > classificationSignature.sign} text file, whereas the custom signatures are read from the \verb{output > signatures > customSignatures.sign} text file.}

\item{distinguishingGenes}{A boolean determining if one or several gene(s) should be used (if set to \code{TRUE}) to further discriminate the patients according to their relative low, intermediate or high expression of this/these gene(s).}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output files are exported to the \code{output > apply} folder. Their precise amounts and organization varies according to the \code{signatureUsed} and \code{distinguishingGenes} arguments (see the online tutorial for more details).
}
\description{
This function computes the Quantitative Scores and the Clinical Scores individually for each patient of a whole dataset, either for the Classification Signature or for one or several custom gene signatures.
}
