% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Cleans the dataset once the related data (RNA-Seq + clinical) are downloaded, matched and processed}
\usage{
cleanData(fullData = fullData, verbose = TRUE)
}
\arguments{
\item{fullData}{A single table with patients in rows and genes in columns.}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output files are exported to the \code{output > data} folder.
}
\description{
This function cleans the dataset after its preprocessing. It is automatically called by the related pooling functions and notably remove columns that are not genes, removes duplicated/incorrect patients, formats gene names to the latest HGNC format, merges duplicated genes, normalizes and transforms the dataset, exports the final dataset and zips the original RNA-Seq and clinical data.
}
