% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetaResults.R
\name{createMetaResults}
\alias{createMetaResults}
\title{Computes the metaResults analysis}
\usage{
createMetaResults(
  selectedGenesOnly = FALSE,
  signaturesMode = FALSE,
  minNumberOfPatientsPerGroup = 5,
  unsollicitedCores = 2,
  iterationsPerCluster = 16,
  genesCutoff = 15,
  verbose = TRUE
)
}
\arguments{
\item{selectedGenesOnly}{A boolean indicating if all the genes (if set to \code{FALSE}) or only a given subset of genes (if set to \code{TRUE}) should be used for the computation. If set to \code{TRUE}, the desired genes to investigate should be contained in a single column table. Please see the online tutorial for more information.}

\item{signaturesMode}{A boolean determining if the analysis should be conducted on an input data table referring to gene signatures enrichments (if set to \code{TRUE}) instead of individual genes (if set to \code{FALSE}). Although very unlikely to be used, this could still be useful if you have access to other metrics than direct gene expression that you want to confront on overall survival.}

\item{minNumberOfPatientsPerGroup}{An integer indicating the minimum number of patients per group to consider. If the number of patients for each group (that is to say, the low, intermediate and high expression of the current gene) is below this threshold, the associated gene is skipped and all the related values generated by the function are replaced by zeros.}

\item{unsollicitedCores}{An integer indicating the number of CPU cores to let free of solicitation. This is useful to impeach the computer to freeze.}

\item{iterationsPerCluster}{An integer indicating the number of iterations (that is to say, random sampling steps) to perform independently for each gene. The higher this value, the more reliable are the results, but the longer the computation time is.}

\item{genesCutoff}{An integer indicating the gene expression quantile to consider as both low/intermediate and intermediate/high thresholds for the construction of the low, intermediate and high expression groups for the current gene. For instance, a value of 15 means that, for each gene, the 15\% most-expressing patients will be in the high expression group, the 15\% least-expressing patients will be in the low expression group, and the remaining (here 70\%) will be in the intermediate expression group.}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output files are exported to the \code{output > metaResults} folder. Their precise amounts and organization varies according to the \code{selectedGenesOnly} and \code{signaturesMode} arguments (see the online tutorial for more details).
}
\description{
This function computes the metaResults analysis, which basically aims to statistically quantify the impact of each gene on the overall survival of patients using the bootstrapping approach.
}
