% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassificationSignature.R
\name{getClassificationSignature}
\alias{getClassificationSignature}
\title{Extract the Classification Signature from the metaResults table}
\usage{
getClassificationSignature(
  GeneScoreThreshold = NULL,
  Gene_SNR_ExpressionThreshold = NULL,
  exportSignature = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{GeneScoreThreshold}{A numeric value indicating the threshold above the one genes are kept relative to their GeneScore.}

\item{Gene_SNR_ExpressionThreshold}{A numeric value indicating the threshold above the one genes are kept relative to their Gene_SNR_Expression.}

\item{exportSignature}{A boolean indicating if the function should actually export the Classification Signature (if set to \code{TRUE}) or not (if set to \code{FALSE}).}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output files are exported to the \code{output > signatures} folder.
}
\description{
This function helps to visually extract the most relevant genes that will be part of the Classification Signature.
}
