% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listProjectsAttributes.R
\name{listProjectsAttributes}
\alias{listProjectsAttributes}
\title{Lists the projects that are downloadable through AMOCATI}
\usage{
listProjectsAttributes()
}
\value{
Outputs its results directly in the console.
}
\description{
This function lists the GDC projects that are currently suitable for downloading and processing through AMOCATI.
}
