% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{Split the dataset into two new subdataset}
\usage{
splitData(datasetAProportion = 0.5, verbose = TRUE)
}
\arguments{
\item{datasetAProportion}{A numeric value indicating the proportion of patients to put into the new subdataset \code{A}. The remaining patients will be put in the new subdataset \code{B}.}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output files are exported to the \code{output > data} folder.
}
\description{
This function splits the dataset into two new subdatasets (\code{A} and \code{B}), using any proportion of interest. This is useful in datasets with a lot of samples to construct intra-cohort training/validation datasets.
}
