% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARGET.createMetaMapping.R
\name{TARGET.createMetaMapping}
\alias{TARGET.createMetaMapping}
\title{Create the metamapping for TARGET projects}
\usage{
TARGET.createMetaMapping(projectID = NULL, verbose = TRUE)
}
\arguments{
\item{projectID}{A character value determining which project to use.}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output file is exported to the \code{output > data} folder.
}
\description{
This function creates the metamapping related to TARGET projects from GDC.
}
