% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TARGET.pool.R
\name{TARGET.pool}
\alias{TARGET.pool}
\title{Pools the RNA-Seq and clinical data from the desired TARGET project}
\usage{
TARGET.pool(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\description{
This function opens and pools the RNA-Seq and clinical data from the desired GDC TARGET project.
}
