% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_CGCI.createMetaMapping.R
\name{TCGA_CGCI.createMetaMapping}
\alias{TCGA_CGCI.createMetaMapping}
\title{Create the metamapping for TCGA or CGCI projects}
\usage{
TCGA_CGCI.createMetaMapping(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output file is exported to the \code{output > data} folder.
}
\description{
This function creates the metamapping related to TCGA or CGCI projects from GDC.
}
