% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_CGCI.download.R
\name{TCGA_CGCI.download}
\alias{TCGA_CGCI.download}
\title{Downloads the data (RNA-Seq + clinical) from the desired TCGA or CGCI project}
\usage{
TCGA_CGCI.download(projectID = NULL, verbose = TRUE)
}
\arguments{
\item{projectID}{A character value determining which project to use.}

\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\value{
Generated output files are fist exported to the \code{temp} folder before being moved to the \code{output > data} folder.
}
\description{
This function downloads the data (RNA-Seq + clinical) from the desired GDC TCGA or CGCI project.
}
