% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_CGCI.pool.R
\name{TCGA_CGCI.pool}
\alias{TCGA_CGCI.pool}
\title{Pools the RNA-Seq and clinical data from the desired TCGA or CGCI project}
\usage{
TCGA_CGCI.pool(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean determining if the function should regularly output its progression in the console (if set to \code{TRUE}).}
}
\description{
This function opens and pools the RNA-Seq and clinical data from the desired GDC TCGA or CGCI project.
}
